Move sub-classes of SendKeys class to separate file
